fkdofficial_v6

+++ obr/src/main/java/com/sayone/obr/ui/controller/AdminController.java
@@ -1,18 +1,23 @@
 package com.sayone.obr.ui.controller;
 
 import com.sayone.obr.dto.UserDto;
+import com.sayone.obr.exception.AdminErrorMessages;
 import com.sayone.obr.exception.PublisherErrorMessages;
 import com.sayone.obr.model.request.AdminDetailsRequestModel;
 import com.sayone.obr.model.request.PublisherDetailsRequestModel;
 import com.sayone.obr.model.response.AdminRestModel;
 import com.sayone.obr.model.response.PublisherRestModel;
 import com.sayone.obr.model.response.UserRestModel;
+import com.sayone.obr.repository.BookRepository;
+import com.sayone.obr.service.BookService;
 import com.sayone.obr.service.UserService;
 import org.springframework.beans.BeanUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.web.bind.annotation.*;
+
+import java.util.Objects;
 
 @RestController
 @RequestMapping
@@ -21,10 +26,15 @@ public class AdminController {
     @Autowired
     UserService userService;
 
+    @Autowired
+    BookService bookService;
+
     @GetMapping("/admin/get-publishers")
-    public PublisherRestModel getAllPublishers(String role) {
+    public PublisherRestModel getAllPublishers(String role) throws Exception {
         Authentication auth = SecurityContextHolder.getContext().getAuthentication();
         UserDto user = userService.getUser(auth.getName());
+
+        if (!Objects.equals(user.getRole(), "admin")) throw new Exception(AdminErrorMessages.NO_ADMIN_FOUND.getAdminErrorMessages());
 
         PublisherRestModel returnValue = new PublisherRestModel();
 
@@ -35,9 +45,11 @@ public class AdminController {
     }
 
     @GetMapping("/admin/get-users")
-    public UserRestModel getAllUsers() {
+    public UserRestModel getAllUsers() throws Exception {
         Authentication auth = SecurityContextHolder.getContext().getAuthentication();
         UserDto user = userService.getUser(auth.getName());
+
+        if (!Objects.equals(user.getRole(), "admin")) throw new Exception(AdminErrorMessages.NO_ADMIN_FOUND.getAdminErrorMessages());
 
         UserRestModel returnValue = new UserRestModel();
 
@@ -81,20 +93,35 @@ public class AdminController {
 //    }
 
     @DeleteMapping("/admin/del/publisher/{id}")
-    public void deletePublisher(@PathVariable String id) {
+    public void deletePublisher(@PathVariable String id) throws Exception {
 
         Authentication auth = SecurityContextHolder.getContext().getAuthentication();
         UserDto user = userService.getUser(auth.getName());
+
+        if (!Objects.equals(user.getRole(), "admin")) throw new Exception(AdminErrorMessages.NO_ADMIN_FOUND.getAdminErrorMessages());
 
         userService.deletePublisher(id);
     }
 
     @DeleteMapping("/admin/del/user/{id}")
-    public void deleteUser(@PathVariable String id) {
+    public void deleteUser(@PathVariable String id) throws Exception {
 
         Authentication auth = SecurityContextHolder.getContext().getAuthentication();
         UserDto user = userService.getUser(auth.getName());
 
+        if (!Objects.equals(user.getRole(), "admin")) throw new Exception(AdminErrorMessages.NO_ADMIN_FOUND.getAdminErrorMessages());
+
         userService.deleteUser(id);
+    }
+
+    @DeleteMapping("/admin/del/book/{id}")
+    public void deletePostedBook(@PathVariable Long id) throws Exception {
+
+        Authentication auth = SecurityContextHolder.getContext().getAuthentication();
+        UserDto user = userService.getUser(auth.getName());
+
+        if (!Objects.equals(user.getRole(), "admin")) throw new Exception(AdminErrorMessages.NO_ADMIN_FOUND.getAdminErrorMessages());
+
+        bookService.deletePostedBookByAdmin(id);
     }
 }
